Refactor create_task.js & add more daily options
Refactor create_task.js into a folder with seperate file for
  dailies, habits,todos, and rewards for easier readability

Add support for monthlies and yearlies.

Additionally fix weeklies's repeat on days bug were it didn't set them, and various spelling mistakes.
