+++ creates/create_task/input_fields/reward.js
@@ -1,0 +1,9 @@
+module.exports = (z, bundle) => [
+    {
+        key: 'value',
+        label: 'Price',
+        helpText: 'Only applicable to Rewards. Set the price of the reward.',
+        type: 'number',
+        required: false,
+    }
+];
