+++ creates/create_task/input_fields/todo.js
@@ -1,0 +1,9 @@
+module.exports = (z, bundle) => [
+    {
+        key: 'date',
+        label: 'Due Date',
+        helpText: 'Applicable only to To-Dos. Choose the date you want the To-Do to be due.',
+        type: 'datetime',
+        required: false,
+    }
+];
